+++ src/vistas/inicio.js
@@ -1,7 +1,10 @@
-export default () => {
-    const shows = `<h1>Bienvenidos</h1>
-<p>Esta es nuestro inicio de proyecto</p>`
-    const divElement = document.createElement('div');
-    divElement.innerHTML = shows;
-    return divElement;
+const showFirst = () => {
+    let elementContainer= document.getElementById('container');
+elementContainer.innerHTML =
+`<div class="registro" id="registro">
+
+</div>`
+;
 };
+
+export default showFirst;
