--- webpack.config.js
@@ -1,25 +1,0 @@
-const HTMLWebpackPlugin = require('html-webpack-plugin')
-module.exports = {
-    entry: './src/main.js',
-    output: {
-        path: __dirname + '/dist',
-        filename: 'bundle.js'
-    },
-    module: {
-        rules: [
-            {
-                test: /\.css$/i,
-                use: ['style-loader', 'css-loader']
-            },
-            {
-                test: /\.html$/i,
-                loader: 'html-loader',
-            }
-        ]
-    },
-    plugins: [
-        new HTMLWebpackPlugin({
-            template: './src/index.html'
-        })
-    ]
-}
